Refactor: move OperationRepository and Operation to domain.core
This commit moves the `OperationRepository` interface and the `Operation` data class from the `domain.game` module to the `domain.core` module.

**Key Changes:**

-   **`OperationRepository` and `Operation` Moved:**
    -   The `OperationRepository.kt` file (containing the interface and data class) has been deleted from `domain/game/src/main/java/com/example/domain/game/repositories/` and created in `domain/core/src/main/java/com/example/domain/core/`.
-   **Import Updates:**
    -   All usages of `OperationRepository` and `Operation` have been updated to import from `com.example.domain.core` instead of `com.example.domain.game.repositories`. This affects:
        -   `data/game/src/main/java/com/example/data/game/mappers/ProtoOperationMapper.kt`
        -   `data/game/src/main/java/com/example/data/game/AndroidProtoOperationRepository.kt`
        -   `domain/game/src/main/java/com/example/domain/game/usecases/UndoOperationUseCase.kt`
        -   `domain/game/src/main/java/com/example/domain/game/usecases/RedoOperationUseCase.kt`
        -   `domain/game/src/main/java/com/example/domain/game/usecases/ResetGameUseCase.kt`
        -   `feature/game/src/main/java/com/example/feature/game/SudokuGameViewModel.kt`
        -   Test files: `RedoOperationUseCaseTest.kt` and `UndoOperationUseCaseTest.kt` in `domain/game`.
-   **Dependency Injection:**
    -   `data/game/src/main/java/com/example/data/game/GameModule.kt`: The module now correctly provides `com.example.domain.core.OperationRepository`.
    -   `domain/creator/src/main/java/com/example/domain/creator/Module.kt`: `CreateNewGameUseCase` now correctly depends on `com.example.domain.core.OperationRepository`.

**Rationale:**

The `OperationRepository` and `Operation` data class are related to the core concept of undo/redo operations which can be potentially used by different parts of the domain, not just the game playing feature. Moving them to `domain.core` makes them more broadly accessible and aligns with better domain-driven design principles. Clearing operations upon new game creation and game completion ensures a clean state for the undo/redo stack.
