+++ data/game/src/main/java/com/example/data/game/GameModule.kt
@@ -3,8 +3,8 @@ package com.example.data.game
 import com.example.data.core.proto.ProtoStorage
 import com.example.data.core.proto.ProtoStorageFactory
 import com.example.domain.core.GameRepository
+import com.example.domain.core.OperationRepository
 import com.example.domain.game.repositories.GameResultWriter
-import com.example.domain.game.repositories.OperationRepository
 import data.game.ProtoGame
 import org.koin.dsl.module
